+++ src/classes/babies/PrettyBaby.ts
@@ -1,7 +1,7 @@
-import type { GridEntityXMLType } from "isaac-typescript-definitions";
 import {
   EntityType,
   FireplaceVariant,
+  GridEntityXMLType,
   ItemPoolType,
   LevelStage,
   ModCallback,
@@ -74,7 +74,7 @@ export class PrettyBaby extends Baby {
     const roomType = room.GetType();
 
     if (shouldTransformRoomType(roomType)) {
-      return [999, 0, 0]; // Equal to 1000.0, which is a blank effect, which is essentially nothing.
+      return [GridEntityXMLType.EFFECT, 0, 0];
     }
 
     return undefined;